mail/requirements.txt
yagmail==0.15.293
pyzmail36==1.0.5
imapclient==2.3.1
requests==2.31.0
python-dotenv==1.0.0 